Improved efficiency in DebugLinesLayer.
The BIG problem with the way DebugLinesLayer worked is that it kept ALL
the lines it needed to draw internally (not a problem) and then drew
them every draw() call (big problem).  The answer is to create a
RenderTexture and draw to this instead.  Add it as a layer to the scene
and erase the lines by clearing the texture when needed.

TRADEOFF:  Costs memory to hold a render texture for the whole screen
vs. minimal hit to performance for redrawing the entire texture with
blending (negligble).