Width generation response improved.
I added a feedback loop (proportional only) around the width and it has
something close to the desired response.  When the difference is small
between the estimated width and the current width, it changes very
little or not at all.  When the difference is great (i.e. fast stroke),
the width grows quickly.  With the current settings, a fast swipe has
an estimate for the width around 14.5.  I set the upper limit to 10 so
that it could "max out" when doing fast draws, but still handle slow
speeds well.  Some tweaking necessary, but path looks promising.