Baseline Solution
There are still some small nuances to work out, like how to make the
coloring work properly when two colors overlap (this is a function of
the blending mode).  There is also a section in the book by Philip
Rideout where he uses an image for the line segments instead of
triangles, which might work out better (though handling multiple widths
will be a challenge).  I think its in good enough shape for the
tutorial pages to be put together, regardless, so this will be the
baseline.